pyproject.toml
[tool.poetry]
name = "ci-agent"
version = "0.1.0"
description = "AI Agent for collecting competitive intelligence."
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
black = "24.10.0"
pylint = "3.3.3"
rapidfuzz = "3.11.0"
edgartools = "3.5.1"
python-dotenv = "1.0.1"
openai = "1.60.2"
sentence-transformers = "3.3.1"
pandas = "2.2.3"
fastapi = "0.115.7"


[tool.poetry.group.dev.dependencies]
gradio = "5.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
